initial work to prototype distributed caching

+++ src/Maw.Domain/Blogs/IBlogService.cs
@@ -7,8 +7,8 @@ namespace Maw.Domain.Blogs
 	public interface IBlogService
     {
         Task<IEnumerable<Blog>> GetBlogsAsync();
-        Task<IEnumerable<Post>> GetAllPostsAsync(byte blogId);
-        Task<IEnumerable<Post>> GetLatestPostsAsync(byte blogId, short postCount);
+        Task<IEnumerable<Post>> GetAllPostsAsync(short blogId);
+        Task<IEnumerable<Post>> GetLatestPostsAsync(short blogId, short postCount);
         Task AddPostAsync(Post post);
     }
 }
